requirements.txt
Pillow>=9.0.0
pyautogui>=0.9.53
pyinstaller>=5.7.0
keyboard>=0.13.5 